requirements.txt
questionary
PyGithub
PyYAML
requests
zipfile